+++ lihil/utils/typing.py
@@ -1,5 +1,14 @@
 from types import GenericAlias, UnionType
-from typing import Annotated, Any, TypeAliasType, Union, cast, get_args, get_origin
+from typing import (
+    Annotated,
+    Any,
+    Sequence,
+    TypeAliasType,
+    Union,
+    cast,
+    get_args,
+    get_origin,
+)
 
 
 def flatten_annotated[T](
@@ -28,5 +37,15 @@ def is_union_type(t: type | UnionType | GenericAlias | TypeAliasType):
     return get_origin(t) in (Union, UnionType)
 
 
-#def isasyncfunc():
+def is_nontextual_sequence(type_: Any):
+    if not isinstance(type_, type):
+        return False
+
+    if type_ in (str, bytes):
+        return False
+
+    return issubclass(type_, Sequence)
+
+
+# def isasyncfunc():
 #    return iscoroutinefunction