+++ lihil/vendor_types.py
@@ -3,3 +3,4 @@ from starlette.responses import Response as Response
 from starlette.types import Lifespan as Lifespan
 
 from starlette.datastructures import FormData as FormData
+from starlette.datastructures import UploadFile as UploadFile